server/requirements.txt
flask
pymongo
python-dotenv
flask-cors
bcrypt
pyjwt
requests
